clarify versions to use for netstandard libraries
Is this change correct? It looks to me like the nuget packages 4.3.x are the first to support netstandard2.x and 4.2x are the first to support netstandard1.6. Should libraries target the highest or lowest third digit? For example, for a library targeting netstandard2.0, should it target FSharp.Core nupkg or 4.3.1 or 4.3.4?